+++ src/main/java/mejai/mejaigg/match/matchdatestreak/repository/MatchDateStreakRepository.java
@@ -1,4 +1,4 @@
-package mejai.mejaigg.matchdatestreak.repository;
+package mejai.mejaigg.match.matchdatestreak.repository;
 
 import java.sql.Date;
 import java.util.Optional;
@@ -6,7 +6,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import mejai.mejaigg.matchdatestreak.entity.MatchDateStreak;
+import mejai.mejaigg.match.matchdatestreak.entity.MatchDateStreak;
 
 public interface MatchDateStreakRepository extends JpaRepository<MatchDateStreak, Long> {
 	@Query("select mds from MatchDateStreak mds where mds.date = :date "
