+++ src/main/java/mejai/mejaigg/searchhistory/entity/SearchHistory.java
@@ -12,7 +12,7 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import lombok.Getter;
 import mejai.mejaigg.global.jpa.BaseEntity;
-import mejai.mejaigg.matchdatestreak.entity.MatchDateStreak;
+import mejai.mejaigg.match.matchdatestreak.entity.MatchDateStreak;
 import mejai.mejaigg.summoner.entity.User;
 
 @Entity
