+++ src/main/java/mejai/mejaigg/summoner/service/UserStreakService.java
@@ -17,9 +17,9 @@ import org.springframework.web.server.ResponseStatusException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import mejai.mejaigg.global.util.YearMonthToEpochUtil;
+import mejai.mejaigg.match.matchdatestreak.entity.MatchDateStreak;
+import mejai.mejaigg.match.matchdatestreak.repository.MatchDateStreakRepository;
 import mejai.mejaigg.match.repository.MatchRepository;
-import mejai.mejaigg.matchdatestreak.entity.MatchDateStreak;
-import mejai.mejaigg.matchdatestreak.repository.MatchDateStreakRepository;
 import mejai.mejaigg.riot.service.RiotService;
 import mejai.mejaigg.searchhistory.entity.SearchHistory;
 import mejai.mejaigg.searchhistory.repository.SearchHistoryRepository;
